Fix some tests so that they can be ran without having a specific localization setting in the host. Take into account the last analyse errors of SonarQube.

+++ core-library/src/main/java/org/silverpeas/core/io/media/video/ffmpeg/FFmpegThumbnailExtractor.java
@@ -59,7 +59,7 @@ public class FFmpegThumbnailExtractor implements VideoThumbnailExtractor {
       for (ThumbnailPeriod thumbPeriod : ThumbnailPeriod.ALL_VALIDS) {
         double timePeriod = thumbPeriod.getPercent() * duration.getTimeAsLong() / 1000;
         FFmpegUtil.extractVideoThumbnail(video,
-            new File(thumbnailDir, thumbPeriod.getFilename()), (int) timePeriod);
+            new File(thumbnailDir, thumbPeriod.getFilename()), timePeriod);
       }
     } else {
       SilverLogger.getLogger(this)
